requirements.txt
scikit-learn>=1.3
umap-learn>=0.5
matplotlib>=3.7
numpy>=1.24
